Changing behavior of majiq tsv output to match voila tsv
TSV output of majiq tsv and voila tsv includes columns named "Junctions
coords" and "IR coords". Given identical naming, one could expect their
content to be the same. However, this was not the case. This update
updates the intended behavior of these columns from majiq tsv output to
match the behavior of voila tsv. That is, the length of "junctions
coords" will match the length of quantifications by including
intron coordinates (in addition to junctions), which are always listed
last. "IR coords" will match the last pair of coordinates in "Junctions
coords" when an LSV includes intron retention and will be blank
otherwise.