Check that junction is not exitron before marking as constitutive
Added check that a junction is not an exitron (source/donor exons are
the same) before marking as constitutive. While exitrons may be the only
outgoing/incoming junctions coming into an isolated exon, they are not
valid constitutive junctions because majiq does not attempt to evaluate
inclusion of the associated intron. Thus, majiq is unable to call
exitrons as constitutive