Keep intron coordinates consistent in presence of exon extension
In `Gene::detect_introns()` and `Gene::create_annot_intron()`, intron
coordinates are defined to be offset by one from adjacent exons so that
they do not intersect. That is, an intron between two exons `[a, b]` and
`[c, d]` will have coordinates `[b+1, c-1]`. However, when
`Gene::connect_introns()` is run to handle changed exon boundaries due
to exon extension, this offset is not applied. This commit adds this
same offset to `Gene::connect_introns()`.