Reset counts of experiments passing group filters with sj files
The function `majiq.src.build._parse_junction_file` only calls junction
or intron `clear_nreads(bool reset_grp)` when the parsed junction file
contains the junction. While this is efficient for not resetting reads
that were not seen, this causes junctions that are not seen in the last
sample in each group to not be reset between groups. This leads counts
of samples passing group filters to be preserved between groups,
allowing junctions and introns to pass group filters on the basis of
pooled counts.

This commit implements a simple solution -- if the last sample in a
group is a junction file (and thus `_parse_junction_file` is called with
`reset=True`), we iterate over all junctions/introns per gene and
specifically reset these flags.

I did not evaluate whether the bug exists for input bam files, and this
commit does not address anything but when the last sample in a group is
a junction file.