Fixed functions using saved log prior to use saved absolute prior
Only `generate_high_probability_non_changing` was using saved prior. It
was previously computing incorrect values due to the incorrectly
computed prior in logspace. It assumed that the prior was in logspace.
Fixing how the prior was computed by changing the prior to be in
absolute probabailities rather than log probabilities (to be consistent
with the rest of voila) introduced a regression in this function, which
is addressed by this commit