Fix distribution mean for per-position coverage when normalizing stacks
Goal of normalizing stacks is to compare the number of reads for a
position to a distribution inferred from the remaining position. We
typically asssume that a nonzero position is roughly distributed as a
Poisson/negative-binomial with mean equal to the average over the other
nonzero positions.

Previously, we were incorrectly calculating `const float mean_reads` as
the leave-one-out sum total of reads, rather than the leave-one-out mean
number of reads. This commit correctly divides this sum by the number of
other positions. It also adds an easier check for if the remaining
number of reads is zero -- this only happens when the number of nonzero
positions is exactly 1.