Normalize intron coverage using uniform floating point scaling factor
Previously, intron coverage was normalized per bin with respect to the
number of positions that are assigned to that bin. This is only
consistent in the limit of infinite reads or infinite positions.
Otherwise, we can realize biases from non-uniform coverage favoring bins
scaled up or down compared to a uniform scaling factor