Fixed off-by-one error in region length function
Region intervals are defined in 1-based coordinates by MAJIQ. This
means that intervals are closed on both ends. Previously we were
calculating length as if we had half-closed intervals (for example as
with 0-based coordinates `(a, b]` length is `b - a`). However, an
interval `[a, b]` has length `b - a + 1`.

The length function is currently used in two contexts:

1. for comparing splice sites/junctions
2. for intron length calculations (`Intron::is_reliable()` and
   `Intron::add_read_rates_buff()`)

In the first context, the change has no impact.
In the second context, the change corrects normalilzation of intron read
rates and allows length 1 (previously miscalculated as length 0) to be
considered reliable.