Update how position offsets calculated for junctions/introns
Previous implementation caused junction to be erroneously counted/not
counted depending on genomic position of read relative to junction
start. Update ensures that junctions counted if and only if split in
read occurs in middle of read (not within MIN_BP_OVERLAP positions from
the end of the read). Soft and hard-clipped bases will be ignored in
read length/overlap in considering position on read.