Added functions to parse first 100 lines of BAM to update eff_len
If max_readlen is misspecified, we get biases in intron coverage (too
high) or segmentation faults (too low). These commits enable automatic
parsing of the first 100 reads of a BAM file to increase the effective
length based on what was seen in those first 100 reads. Over 16k BAM
files (including all of GTEx v8) had a read with the maximum length in
the first 31 reads.

We still need to provide a mechanism for updating eff_len and resizing
any buffers that have already been created in case our previous
observations no longer hold. Once that is done, we could potentially
remove the need of specifying a read length whatsoever.