Added safety factor on buff_len vs updated eff_len_
The idea is we want to avoid needint to expand all the buffers multiple
times. If multithreaded, we still need to add locks. However, we
currently are not in this part of the code, which means we could lazily
expand buffers as necessary/let std::vector handle extra buffer for us,
saving space and time.