Resizing buffers only happens when parsing BAMs without concurrrent access
Worry with function to update eff_len_ and buffers is that we might be
resizing the buffers while some other function is updating. It turns out
that we only call this function in contexts where the only
multithreading is for I/O reading a BAM file from disk. Thus, we do not
need to add any locks, and we can in fact parallelize the resizing of
buffers