Fix calculation of leave-one-out-mean of the position coverage
The calculation of the leave-one-out-mean of position coverage was only
correct until the first stack was removed (if any). This lowered the
denominator, increasing the leave-one-out-mean for all subsequent
positions, inflating the threshold required for calling subsequent
stacks. This commit keeps track of this shared denominator over all
positions.

A potential question this raises is whether we would want to repeat the
stack removal process iteratively. Each time a stack is removed, it
lowers all other leave-one-out means. This means that positions that
previously were not outliers would be outliers if we checked again.

I describe it iteratively because it's easiest to think of it in such a
manner with respect to the current implementation. However,
making it so that no positions are outliers with respect to the final
leave-one-out means instead of the original leave-one-out means can be
done with a single pass over sorted data (it would be close to the
implementation before this commit except that we would update the value
of `sreads` as well as `npos`). This would be a trivial cost in time for
a potentially cleaner solution.