Only test positions with most extreme coverage when normalizing stacks
If there are 50 positions with nonzero coverage, we know that if the
maximum value is not a stack, the other 49 values are not a stack. We
take the time to sort the coverage vector under the
assumption/observation that this is less expensive than testing multiple
positions as outliers (requiring calculation of special functions,
especially if using negative binomial) and that most junctions/introns
do not have any outliers.

Since this sorts the coverage vector, it is important that the
bootstrapping that takes place on the non-stack positions is random
(i.e. the changes from PR#37 implementing static random state for
bootstrapping)

The new implementation also makes it trivial to consider removing stacks
that are outliers when considering the distribution of coverage after
each stack is removed. Right now we are looking at the distribution
before any outliers removed. This could make a difference if we have
outliers at different scales, where removing the first outlier could
show additional outliers (i.e. 100 positions
`([20] * 98) + [2000, 200000]`. The 2000 would not be called as an
outlier by the current approach because the effect of 200000 is so
significant on the leave-one-out distribution mean.