Fixed bug where junction position was occasionaly misidentified
Branch fix-junction-positions was made to handle cases where a junction
had multiple reads associated with it. While it improved the condition
of when to add an offset, the updated logic for how to advance the read
and genomic offsets would occasionally fail. This happened when one of
the junctions was before MIN_BP_OVERLAP. In this case, the logic of
adding a junction was ignored by a continue statement. However, the
continue statement ignored the need to still update the genomic offset
as a consequence of the junction. This commit ensures that the genomic
offset is properly updated even in this case by changing the conditional
logic.