Exon count in LSV::set_type is in strand order, retaining junction order
LSV junctions are sorted in strand order with respect to reference exon
coordinate followed by other exon coordinate. An exon count from 1 to
the number of unique other exons is set as part of lsv type for
visualization in VOILA. This count was previously in order of junctions,
yielding errors when the first junction went to a later exon. This
commit pre-computes the exon counts using an ordered set beforehand,
allowing exon counts in the lsv type to be correct without modifying the
order of junctions in the LSV.