Replace scythestat calls for beta cdf with boost calls
Boost library is around 5x faster on laptop, speeding up all our
quantifiers as implemented. On benchmarking the relative error is
negligible except for probabilities less than 1e-12, where scythestat
truncates down to zero, which is less correct. Right now the speedup is
with full default precision, this could be adapted for giving a less
precise answer in exchange for an additional 2-3x speedup (most relevant
for psi and het where the discretized distribution will only be used for
visualization)