Close open intron in detect_introns() even if no room for intron
We add potential introns in the space between open/close splice sites.
This failed on the edge case where open/close splice sites gives us a
zero length intron. So, if another acceptor came later, we would create
an intron intersecting the exon. This bug is also present in master.