Bugfix to `--dump-coverage` when read length greater than specified
Update for `--dump-coverage` was created before branch auto_readlen was
merged. The appropriate variable for specifying position coverage array
size changed names with auto_readlen, which was not changed in the new
code added by dump-coverage. This led to incorrect behavior, including
segmentation faults, when the local effective read length of a file was
different than what was specified (or not specified at all) in input
configuration file. This commit corrects this behavior.