Note that junction/intron positions calculated relative to alignment
Previously, we calculated positions against genomic coordinates, which
was problematic due to the possibility of reads with multiple junctions
leading to invalid offsets. To mitigate this, we switched to using
read position after adjusting for clipping. We note that we can and
should distinguish between read positions before and after adjusting for
clipping. It makes sense to distinguish the read/query as independent of
any alignment, so read positions should be before soft clipping. The
previous read positions calculated after adjusting for clipping can be
distinguished as alignment positions/offsets. This commit renames
variables to distinguish alignment offsets accordingly.

Note that read positions are still counted relative to alignment. This
commit is the first step toward switching to counting relative to
read/query position/offsets while applying criteria for overlap relative
to alignments.