Include minpos in filters required for denovo junction detection
Currently, the only filter being applied for potential denovo junctions
is a minimum number of reads. This can sometimes be weaker than the
requirement for a junction to be reliable, which relies on a minimum
number of reads *and minimum number of nonzero positions*.

In particular, if enough experiments in a group have a single position
with many reads (i.e. a lone stack), we will add it to the splicegraph
as a denovo junction even though we don't believe that any of the
experiments have any coverage.

Instead, we now require a potential denovo junction to meet the same
minimum number of positions filter as the builder so that as long as
`min_denovo >= minreads` an accepted denovo junction will have also
passed build filters.