Only update junction build/denovo flags that need to be updated
Functionally equivalent rewrite of `Junction::update_flags`, with added
comments.
Private variables `flter_cnt_`, `denovo_cnt_` only incremented for
junctions that have not met criteria for `bld_fltr_` and `denovo_bl_`.
Thus, they remain 0 for junctions that have already passed, which saves
time. This is different than before, but these variables are only used
for that purpose.
We also only mutate `bld_fltr_` and `denovo_bl_` when we are going to
change it.
The only potential difference in functionality is an old condition to
set `denovo_bl_` to False. This only happens if `update_flags()` is
called with `denovo=True` so that we pass a denovo junction, followed by
calling it again with `denovo=False`. This does not happen in MAJIQ.