Replace implementation of Wilcoxon for performance and accuracy
Previous Wilcoxon implementation had above quartic (N^4) complexity with
increasing sample size. In practical implementations of the general
statistical test, there is a switch between the exact p-value
calculation to asymptotic approximation using a normal distribution with
identical mean and variance due to the expense of calculating the
statistic. Our previous implementation tried to calculate the exact
p-value even for large samples where it would take many orders of
magnitude more time than the other statistics, leading to unacceptable
runtimes of MAJIQ HET. Additionally, the validity of the implementation
was uncertain due to lack of documentation of what was being calculated.

To address performance/validity, this commit reimplements the
independent two-sample test (typically called the Mann-Whitney U test,
not Wilcoxon, which is for paired samples) from scratch. We use fast,
exact 64-bit integer arithmetic to calculate exact p-values for small
sample sizes. We use normal approximation with continuity correction
otherwise. Caching uses symmetries of the Wilcoxon distribution to
reduce the amount of memory/computation needed. On benchmarking, we
found that this is over 100x faster than the previous implementation for
maximum sample size before switch to asymptotic approximation
(n1+n2=64).