Don't give majiq stats missing data, handle beforehand
We currently copy data around a bunch of times, which is probably
overkill. MAJIQ puts in missing values by setting quantifications to -1,
but some of the tests look for missing values by looking for negative
labels. We can simplify this by just not adding these quantifications to
the lists provided to the statistics functions.
To handle crashes increased by this change, also make tTest return
pvalue 1 when there are insufficient experiments