Use std::pair<int, int> instead of string for coordinate keys
We were creating maps of string to junction/intron/exon objects. We were
iterating over these objects under the assumption that they were sorted
by coordinate. Locally, they are almost always iterable in coordinate
order because lexicographical string order of "{start}-{end}" is very
close to lexicographical pair order of (start, end) as a pair of
integers. The exception when these containers of coordinates cross a
power of 10. For example,
`(900, 1100) < (1300, 1500)` but `"900-1100" > "1300-1500"`.
The most notable instance of where we were doing this was in the
procedure where we detect and connect introns to adjacent exons.
Incorrect intron order led to mismatched exons or undetected introns.
This commit replaces the type used in these map to enforce the correct
sorted order for these purposes.