Fix detection of long alt 3' exon extension
We expect to detect exon extension events up to some distance
MAX_DENOVO_DIFFERENCE. When we have many denovo junctions, we can have a
sequence of denovo junctions that lead to exon extension that should
extend past MAX_DENOVO_DIFFERENCE. For alt 3' exon extension (extension
of exon backwards), we would occasionally incorrectly extend this
backwards. This was because of an error in `Gene::detect_exons()`. We
kept splice acceptors "opening an exon" in a vector. This was used
correctly as a stack in parts of the code, but if we had leftover
elements, we would iterate over the object in the wrong order, making
that earlier junctions would not detect exon extension backwards, making
them half exons incorrectly. This commit corrects the order in which we
iterate over the object, allowing us to detect exon extension correctly.