Fix search of introns to assign coverage
We were incorrectly searching for introns to assign coverage to when
parsing BAM files for intron coverage. We used lower_bound on the vector
of potential introns per contig. This vector of potential introns per
contig was sorted in coordinate order -- first by start, and then by
end. However `std::lower_bound` along with the predicate function used
assumed appropriate partitioning of intron ends (e.g. sufficient
condition: sorted order for the intron ends). This was not true. As a
result, the binary search (performed by lower_bound) would skip introns
that should have been considered. Performing binary search instead of
brute-force makes a significant difference in runtime for intron
retention detection.

This commit creates an auxillary variable with the cumulative maximum of
intron ends (to match the lists of introns in sorted order). This allows
us to appropriately partition the list to acquire the relevant subset of
introns per read efficiently.

This commit also adds needed documentation of the overGene class as an
aggregator of directly/indirectly overlapping genes. This helps explain
why we do not need an equivalent auxillary variable for the several
searches over overGene vectors -- they are appropriately partitioned by
construction.