Update bootstrapping procedure to faithfully model variance
This update removes a fixed value of k from the bootstrapping procedure.
Instead, k is set to exactly one less than the number of nonzero
positions. In addition to removing a user option that was likely
difficult to understand why/how you would change it, the fixed value
caused us to underestimate variance for low coverage junctions and
overestimate variance for high coverage junctions. The underlying
intuition comes from the observation that
`Var(sum(X[i])) = sum(Var(X[i]))` but `Var(a*X) = a^2 Var(X)`.
This applies to how we would rescale a sum over `K` positions to be like
one over `npos` positions. This caused the resulting bootstrap
distribution to have variance scaled incorrectly by a factor of around
`npos / K`. We can mathematically prove that `K = npos - 1` yields a
bootstrap distribution that matches a generating readrate distribution
in mean and variance (in terms of distance between two samples).