Use parametric bootstrap when nonparametric bootstrap underdispersed
We expect our count data to be Poisson or negative binomial (variance >=
mean). Our approach performing bootstrap sampling over nonzero positions
rather than all positions leads to underdisperison in some cases. This
is an underestimate of the true variance under reasonable assumptions on
the data. Thus, in these cases, we now sample from a Poisson
distribution parameterized by the data as a less severe underestimate of
the true variance that will be propagated into quantification