Silence warnings for using numpy deprecated API
We are currently using a really old C API for numpy. This leads to
several build warnings. We are aware of this issue and will need to be
careful with accepting newer versions of numpy in the future if they
remove backwards compatibility. However, this is a warning that we are
aware of and can safely ignore for now.