Do not count junctions sharing boundary of annotated zero-length introns
This changes a conditionals in parse_read_for_ir() that caused
reads with junctions sharing an annotated zero-length intron boundary to
be counted for the zero-length intron erroneously. This change treats
introns slightly differently even for introns with annotated positive
length.

Previously, if we had an intron `[a, b]` (preceded by exon
`[..., a - 1]` and followed by exon `[b + 1, ...]`, we would count reads
with appropriate read offsets that overlapped the intron so long as
there wasn't a junction that would remove bases from the intron in the
resulting transcript. This allowed junctions that ended at `a` or
started at `b`. While such junctions do not remove any bases from the
intronic region itself, it means that the transcript from which this
read originated used a junction to get from one exon to the next. This
commit does not allow these junctions anymore.