Condition for simplification: not enough experiments above threshold
Previously condition for simplification was enough experiments below
threshold. This led to nonintuitive behavior (as determined during group
meeting) when min-experiments was anything other than 50%. In the most
extreme case, if min-experiments was 1, a single experiment with low PSI
would be enough to cause a junction to be considered simplified for that
group -- so all junctions in an LSV could actually be made to be
simplified by the experiments in that build group.

Instead, we now use min-experiments as a threshold on the number of
experiments where PSI is above the threshold (criteria against
simplification).