Defined new `simplifier_min_experiments` variable for simplification
New `simplifier_min_experiments` variable is used in place of usual
`min_experiments` to specify the number of experiments that must pass
the simplification threshold. This commit splits the logic used by MAJIQ
but does not provide a way of changing the value to be different than
old `min_experiments` variable.