Correctly yield multi-HET column names for nonchanging groups
Correctly yield multi-HET column names for nonchanging groups. In the
case of a single comparison, we should yield "nonchanging" without group
information. However, we incorrectly returned "nonchanging" instead,
which upon iteration would yield each character in "nonchanging" one at
a time. This commit fixes this. It is a wonder that this worked in the
first place.