Compute p-values for posterior means and psisamples
We expect p-values from posterior means to be better calibrated
than some quantile from samples of the posterior distributions.
We still want such quantiles for reproducibility. This commit
updates the psisamples structure/how it is used to have an
additional column per junction (the first) devoted for the posterior
means. Then, pvalues are computed on the posterior means as well as
any psi samples.