Reimplemented GFF3 parsing to handle more general hierarchies
GFF3 parsing was predicated on 3 feature levels based on our experience
with Ensembl annotations. This is not followed in Refseq annotations, so
this update adds more general way of following hierarchy (specified by
ID and Parent attributes to follow parents of exons (potential
transcripts) to the first ancestor that we accept as a gene. This is
read into a separate structure that encodes our usual model of
transcripts (gene->transcripts->exons), which can be converted into a
splicegraph.