implement approach to assign passed junctions
And realized that this isn't what we've been doing with MAJIQ so far in
the past. When junctions are unstranded, we use annotations to help
infer status. By splitting these junctions into +/- strands at the group
level, we are losing this ability, and subsequently assigning to
lower-evidence genes on the opposite strand.

Thankfully, a lot of current approach still applies to new solution we
need to implement. I sketch the general idea below:

+ pointer to base junctions for search, etc.
+ vector along base junctions with counts passing build filters
+ container of overgenes (implement so not wasting space, only minimal)
+ map of denovo junctions per overgene (gene order)
+ given denovo junction/overgene update this map
+ option to update annotated junctions in place (min_experiments and all
  that)

+ add to passed junctions container
+ vector of booleans passing build filters instead of counts
+ set of denovo junctions per overgene (gene order) instead of map (easy
  to merge into)
+ make new junctions just by merging the original junctions (modified by
  booleans) and the set of denovo junctions per overgene