Replace hashing of regions with checksums.
Turns out that C++ hash functions differ each time the program starts
for good reasons (avoiding malicious users trying to force hash
collisions to degrade performance). Unfortunately, we were trying to use
this as a checksum, where it needs to always be the same. So, we add
Boost's CRC library to compute CRC32 checksums.

Note that we are applying the generator to individual values. We do not
just apply to full struct because:

1. most of structures have a pointer of one kind or another that doesn't
   necessarily have the same values between runs
2. different systems may have different padding, and even if they are
   the same, the padding is not zeroed out, so a checksum over the full
   struct would consider the uninitialized data too

It is conceivable that we might get different checksums on different
platforms because some of our types are implementation dependent... but
this is a good start, and we can always remove these checks moving
forward