Prevent junctions from being placed outside of the range of exons
Sometimes gene coordinates from GFF3 do not match exons/transcripts that
were actually accepted for inclusion into the actual splicegraph.

At this time, we keep the original coordinates for the gene in the GFF3
but use the extreme coordinates of the exons when deteermining which
denovo junctions can be assigned.
Instead, we add a check that the exon coordinates are subset to the gene
coordinates (as expected from valid annotations).
This check is sufficient for the limited cases in which we use gene
coordinates