Condition for adding factors was too weak, brute force checking rank
Checking that the new factors were independent of the old factors by
least squares and relying on orthonormality of new factors (by SVD) was
incorrect because the residuals after subtracting projection to span of
old factors may no longer be orthogonal.
This is not a slow part of the moccasin pipeline, so we just are
brute-force computing rank on increasingly augmented model matrix to
confirm that it would be full rank.
If we wanted maximum efficiency we could do least squares and then
remove projections of the singular vectors until the next singular
vector was ~ 0 (sort of reverse Gram-Schmidt) but would have to be
careful about numerical stability but this is unnecessary/low priority