Load updated events coverage before saving to avoid zarr error
zarr requires chunks to be uniformly sized, but dask does not. The
dask command xarray uses to replace old bootstraps with corrected
bootstraps can lead to non-uniform chunk sizes. So, saving this directly
to zarr leads to error. It's easy enough to just load the data into
memory before saving it to zarr, letting it handle the chunks