We know that coordinates/dimensions are same across files
Fix loading of offsets vector when loading from large numbers (~500+) of
experiments, where xarray/dask was maybe trying to do some kind of check
that it was equal in all experiments? (it was causing deep recursion
leading to exception after passing Python's maximum recursion depth).
Since these non-data coordinates/dimensions are the same given same
splicegraph, etc. we assume equality and just take these values from the
first file. We are still checking that the sizes match, just not that
the order is the same (when labels are available).