Correctly parse junctions from BAMs without stored query sequences
BAM files are sometimes stored without aligned query sequences for
storage purposes (SEQ = "*"). Junction/intron information is still there
in the CIGAR string. However, the previous definition of read length we
were using is not correct when SEQ = "*" -- htslib sets b_->core.l_qseq
(length of query sequence) to the length of the query sequence and
checks it against the CIGAR string when SEQ is not "*", but it actually
sets it to 0 when SEQ is "*" (rather than computing from CIGAR string).

This commit computes the length from the CIGAR string when
b_->core.l_qseq is 0, allowing us to parse these BAM files properly