Implemented way to project SJBinsReads to different SJJunctions
This new functionality allows us to map SJBinsReads to other
SJJunctions. This is named after the idea of a projection map between
two spaces where coordinates present in both spaces are kept, while
coordinates not present in either are zero (or not there) in the output.
This is modified with the caveat that we can map to ambiguous strands.
In these cases, all possible strands (+, -, .) can map, adding read
counts per bin.

This can be combined with the recent additions to SJJunctions that allow
converting stranded junctions to unstranded and flipping strand