First implementation of auto strand detection in new-majiq sj
This is a first implementation of auto strand detection. It includes
hard-coded values and does not separate itself into its own function. It
automatically detects strand by counting number of junctions mapped in
forward vs reverse strand to junctions in input splicegraph. We do not
ignore junctions that are in splicegraph multiple times or in multiple
directions. We use binomial model with p=0.5 and N=forward+reverse
number mapped to calculate two-sided p-value, and set low pvalue cutoff
to reject unstranded null. If we reject null hypothesis, we pick the
strand that gives us the most mapped junctions.