Update chunking of different majiq data structures in zarr
Splicegraph and SJ files are always read at once because they are
typically not meant to align with anything else of the same type
typically 1-2 splicegraphs at a time at most, and SJ files need to have
their data dumped to match events/connections in splicegraph. As a
result, it is ideal to *not* chunk these files (or rather make a single
chunk). We explicitly specify this chunking before saving to ensure that
zarr defaults don't give us nonideal solution.

We also update PsiCoverage to handle chunking when there are zero
events. This is actually a bug in zarr or xarray, but this should work
for now.