Fix Junction::get_key(gene, strandness), fixing stranded simplify
Junction::get_key(gene, strandness) is supposed to return a string with
format "{gene_id}:{strand:[.-+]}:{start:\d+}-{end:\d+}" to match keys
found in SJ files (formed by a different function). This is only used in
the simplifier.

This function did not create the key properly when an experiment was not
stranded because it used std::to_string(char --> int implicit cast) on
the gene strand. So, instead of getting '+' or '-' for the gene strand
in the key, it was instead getting "43" or "45".

This resulted in simplifier with stranded build identifying zero
junction reads, leading to simplification of all LSVs.

This commit changes how the junction key is formed, resolving issue
where 0 LSVs were found with stranded/simplified builds.

(All Credit: Joseph Aicher)