Always write to zarr with consolidated=True
Xarray v0.19.0 makes writing consolidated zarr the default.
[xarray github issue #5251](https://github.com/pydata/xarray/issues/5251)
describes performance considerations.
We want to take advantage of this performance benefit, and we also don't
want it so that files produced using and older version of xarray raise
warnings when opened up in a newer version of xarray. As a result, we
will always write zarr files as consolidated from now on.