Remove INFOSCORE from default statistics when using HET
TTEST is O(N), WILCOXON/TNOM are O(NlogN), INFOSCORE is O(N^2). We don't
take advantage of O(N) TTEST because we sort data even if running only
TTEST. But INFOSCORE runtime complexity becomes slow when getting into
comparisons between 100s of samples, which is within realm of use cases
we expect. As a result, we are leaving INFOSCORE as an option but
removing it from the tests that are run by default