Use combine_nested instead of combine_by_coords in open_mfdataset
PsiCoverage uses xr.open_mfdataset to open multiple PsiCoverage files in
transparent way. Default is to use combine_by_coords. This function
attempts to make concatenated dimensions sorted (and fails if it can't).
When we are concatenating arrays where we have >1 entry, this isn't
always possible. While it's nice to be sorted, it's definitely not
essential, so we fall back to combine_nested instead.

This is also the case for moccasin factors, so we change that too.