Rewrite nm.SpliceGraphReads, encapsulating MultiSpliceGraphReads
Following example of PsiCoverage, we rewrite nm.SpliceGraphReads to
handle the case of more than one expeirment at a time. This disconnects
it from the internals library SpliceGraphReads object, which is used in
SJExperiment and SimplifierGroup. We provide a private helper function
to access the internal representatio as needed, and change the public
interface of SimplifierGroup to look for SJExperiment instead.