Initial implementation of PsiOutliers class
PsiOutliers is the backend supporting comparison of N~1 cases vs N>>1
controls, looking for outliers in PSI. This commit defines the
container, how quantiles for comparison between controls and cases are
specified, and properties for using PsiCoverage to compute quantiles
that will be compared.

In future work, we will implement the actual metrics evaluating the gaps
between quantiles, other functionality.