Output barebones v2 splicegraph to allow using voila tsv
Previously made legacy-majiq to output v2 majiq files for
quantification. This was enough to test majiq psi, majiq deltapsi, but
since the only way to get tsv output from het is through voila, we
needed to make a matching splicegraph. We have introduced new subcommand
new-majiq legacy-splicegraph to enable this