Implement median over bootstrap means
This is what was reported as bootstrap mean rather than the mean of
bootstrap means for direct comparison of old vs new.
JKA: current view is that we should generally do computation on
raw_psi_mean instead of bootstrap_psi_mean or bootstrap_psi_mean_legacy
(newly implemented) because as num_bootstraps increases towards
infinity, these estimators come very close to raw_psi_mean.
It is not asymptotic, but the error is much smaller than standard error
with default of 30 bootstrap replicates.
But if one wants the legacy computations or similar, we now have it.