Do not request boost for lower precision computation in BetaMixture
We were asking for lower precision in BetaMixture computations of beta
cdf, quantiles in calls to boost with the idea that it would be faster.
On profiling, there are only marginal differences in speed for cdf, and
negligible differences for quantile. More importantly, quantile gives
invalid/divide-by-zero warnings if lower precision is used. As a result,
we are no longer going to ask for any change to calculated precision in
BetaMixture