Enable make_annotated for new junctions
GeneJunctionsAccumulator has argument make_annotated which is expected
to set all input junctions to annotated (i.e. not denovo).
This argument was only respected when the GeneJunction was already
observed in a previously added GeneJunctions.
This commit fixes the behavior so that the junction is marked as
annotated even if it is the first time that the accumulator has seen the
junction.