Enable logging function when loading from GFF3
When loading GFF3 files, there are sometimes exons that are not
explicitly ignored or accepted.
It is of potential interest for improving how the GFF3 file is handled.
This commit introduces a callback into the functions that load GFF3
files, allowing for information about the reasons that these exons were
skipped and how many times they were skipped to be handled arbitrarily.
This is used to log warnings on the command-line application