Do NOT release GIL when working with python/pybind objects
Update to NumPy 1.22 showed segfaults when using functions with
pybind11::vectorize under the hood.
This only takes place in conjunction with releasing the GIL.
In these cases, the calls through pybind probably need the GIL, even
before this updated NumPy version.
This commit removes calls to release GIL when we interact with pybind11
API (arrays, objects, str, and vectorize).

In the reproducible test cases I have available, this works with the
updated version of NumPy, so I have unpinned the version of NumPy again.