Explicitly use ThreadPool in for non-dask multithreading
Previously using multiprocessing.dummy.Pool, which is an alias for
ThreadPool.
Internally, this is not a change.
But, this makes explicitly clear that we are using multithreading.
Previously was misleading because usually it is
`import multiprocessing as mp`, not `import multiprocessing.dummy as mp`
so could easily be misinterpreted as multiprocessing