Remove references in docs to methods that have been encapsulated
DeltaPsi used to provide methods to access relevant information from
PMFSummaries objects.
This has been replaced by making a subclass of PMFSummaries specific for
dPSI (DeltaPsiPMF), and making it easy to access those objects from
DeltaPsi.
The docs hadn't been updated for this change, so this commit updates the
documentation accordingly, removing the warnings when building the docs