Fix handling of zero length introns, add minimum length check to Regions
We have two separate use-cases of IntervalPrecedes where we want
different behavior.
Previous implementation was incorrect for either case.
For Regions subclasses where HAS_OVERLAPS = false (for example, introns,
exons), we want to allow for zero-length regions at the boundaries.
That is around region [a, b] (where b >= a), we wish to permit [a, a-1]
and [b+1, b].
In this situation, we wish to say [a, a-1] precedes [a, b] precedes
[b+1, b].
Meanwhile, in most other cases, we wish to say that these are
overlapping.
For example, an intron [a, b] might be split by a new exon [a, b],
leaving two introns [a, a-1], [b+1, b].
These need to be matched back to [a, b].
This similarly applies for assigning intron coverage.

Previous behavior was unconditionally [a, a-1] does not overlap, but
[b+1, b] does, which does not handle either case.

Also added a check for minimum length in Regions which enforces
assumption that exons, junctions, genes must have length > 0